Added check for user id and moderator status in deleteComment

+++ repository/comments.js
@@ -182,9 +182,17 @@ const updateComment = async (args, context) => {
     }
 }
 
-const deleteComment = async (req, res) => {
-    const postId = parseInt(req.params.postId)
-    const commentId = parseInt(req.params.commentId)
+const deleteComment = async (args, context) => {
+    const { postId,
+            commentId,} = args
+    const { user } = context
+
+    if (!user) {
+        console.log(
+            'Unauthenticated user cannot delete comments'
+        )
+        return null;
+    }
 
     try {
         const comment = await db.Comment.findByPk(commentId)
@@ -197,18 +205,35 @@ const deleteComment = async (req, res) => {
             throw new Error('Comment not found')
         }
 
+        //checking whether user is deleting his own comment or a moderator is deleting someone else's comment
+        const post = await db.Post.findByPk(postId)
+        const { userId,
+                communityId,} = post.toJSON()
+        const userCommunity = await db.userCommunity.findOne({
+            where: {
+                userId,
+                communityId,
+            },
+        })
+
+        if (user.id !== userId && !userCommunity.toJSON().isModerator) {
+            throw new Error(
+                'User cannot delete a comment that is not his own when he is not a moderator'
+            )
+        }
+
         await db.Comment.destroy({
             where: {
                 id: commentId,
             },
         })
 
-        res.status(202).send('Comment deleted successfully')
+        return comment
     } catch (e) {
         console.error('Error:', e.message)
-        res.send({
+        return {
             error: 'Something went wrong',
-        })
+        }
     }
 }
